feat: 여정 삭제 Spring Security 기능 추가
여정이 해당하는 여행 작성자가 아니면 삭제할 수 없도록 설정

+++ src/main/java/kr/co/fastcampus/travel/domain/itinerary/service/ItineraryService.java
@@ -1,10 +1,13 @@
 package kr.co.fastcampus.travel.domain.itinerary.service;
 
 import kr.co.fastcampus.travel.common.exception.EntityNotFoundException;
+import kr.co.fastcampus.travel.common.exception.MemberMismatchException;
 import kr.co.fastcampus.travel.domain.itinerary.entity.Itinerary;
 import kr.co.fastcampus.travel.domain.itinerary.repository.ItineraryRepository;
 import kr.co.fastcampus.travel.domain.itinerary.service.dto.request.update.ItineraryUpdateDto;
 import kr.co.fastcampus.travel.domain.itinerary.service.dto.response.ItineraryDto;
+import kr.co.fastcampus.travel.domain.member.entity.Member;
+import kr.co.fastcampus.travel.domain.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -17,6 +20,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class ItineraryService {
 
     private final ItineraryRepository itineraryRepository;
+    private final MemberService memberService;
 
     @Transactional(readOnly = true)
     public ItineraryDto findItineraryById(Long id) {
@@ -31,8 +35,12 @@ public class ItineraryService {
         return ItineraryDto.from(itinerary);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, String email) {
+        Member member = memberService.findMemberByEmail(email);
         var itinerary = findById(id);
+        if (itinerary.getTrip().getMember() != member) {
+            throw new MemberMismatchException();
+        }
         itineraryRepository.delete(itinerary);
     }
